refactor(runner): split cell.test.ts into focused test files
Split the 5,062-line cell.test.ts into 7 focused files:
- cell-core.test.ts: basic cell operations and utility functions
- cell-proxy.test.ts: createProxy and proxy behavior
- cell-as-cell.test.ts: asCell conversion with and without schemas
- cell-links.test.ts: getAsLink, getAsWriteRedirectLink, getImmutableCell
- cell-hooks.test.ts: toCell and toOpaqueRef hooks
- cell-callbacks.test.ts: success callback tests
- cell-arrays.test.ts: array element conversion
